Finally got the role and topic triggers to work all the way - adding and deleting a role for a person, and adding and deleting a topic associated with a role.
If you add a role to a person's node, the triggers will look in the db tree to see what topics are associated with that role, and then those topics are added to the user's node.  If you add or delete a topic from the /roles/${role} node, the triggers will go find the users that have that role and then add/delete the topic from each of the user nodes.   So these triggers maintain the integrity of the data.

And if you delete a role from a person, the triggers will see what topics are associated with that role and remove those topics from the user's node - so just the opposite of what happens when a role is assigned to a user.

We also have a /roles/${role}/users node that keeps track of what users are associated with each role from a role standpoint.  This node is also automatically kept in sync by the triggers.