THIS IS THE COMMIT WE CARE ABOUT.  Was working on the read_sheets branch earlier on a different maching.  I guess the push/commit never happened on that other machine.  So in this commit, I basically did what was in the other commit.  Plus, I actually got more done in this commit:  I now have spreadsheet data being written to the database the way I like.
And by "the way I like", I mean:  /missions/{mission_id}/sheet_id  and  /missions/{mission_id}/data.

Notice that sheet_id is a literal and the value of the attribute is the sheet's id.  And data is also a literal and the data node has child nodes under it which are integer values starting with 0.  Each of these integer nodes represents a row.  If you expand one of these integer nodes you will see a key/value pair of nodes for each column on that row.