Missions is looking really good now.  When you create a mission, you are sent to the All My Missions screen.  There is now a toggle switch that you can use to toggle the active/not active state of a mission.  This toggle is connected to the Firebase database, so when you toggle a mission on one phone it automatically toggles the state on all other phones.
We have two attributes that we use to denote the active/not active state of a mission.  The first attribute is simply "active" with a boolean true or false value.  The other attribute is "uid_and_active" so that we can query for just a particular user's list of active missions.

Also, we load missions from the top of the screen, unlike how we load new messages in ChatFragment