Calling activity is now being stored under /teams/{team}/activity/all and /teams/{team}/activity/by_phone_number.
We used to store activity just under /activity and we jumbled up raw activity entries with by-phone entries.  The result of that was an "All Activity" screen on the phones that showed a bunch of blank lines because it was trying to treat the nodes keyed by phone numbers the same way it was treating the nodes keyed by auto-generated keys.  The values under these nodes didn't have the same structure, so the phone screens couldn't do anything but display blank lines for the nodes keyed by phone number.