New function: addAttributeToChildren so that we can add an account_disposition attribute to every child of /users.  We do that so that because we changing the query that displays users on the Unassigned Users screen.  We used to just show everyone that was under the /no_roles node.  But that didn't allow us to remove people from that list without granting them access.  So what we did was add another control to the AssignUserVC (iOS) and AssignUserFragment (Android) classes.  This control is a toggle switch that allows us to set the value of this new user attribute called account_disposition.  account_disposition needs to be "enabled" by default.  If someone downloads the app that we don't want to let in, we set account_disposition=disabled.  And the query for the Unassigned Users screen only returns children of /no_roles that have account_disposition=enabled
Note: that account_disposition is actually set on the user's node under /users, not /no_roles.  We have an onWrite trigger in updateUser.js that observes changes to nodes under /users.  And whenever the account_disposition of the user node is found to have changed, we propagate that change to the user's node under /no_roles if that node exists there

New account_status_event: When a user's role is removed, we now write an entry under that user's account_status_events node