userList.js - added some temporary functions (prepended with temp_) which aim to populate all user records with the person's CB ID.  We do this by first setting citizen_builder_id="not populated yet" for all users that don't have a citizen_builder_id attribute (which in the beginning was everyeone).  Then we defined a fuction temp_call_volunteers which is a trigger function.  It is triggered whenever any user is first assigned a citizen_builder_id attribute of "not populated yet".  Since we can't loop through all the users, the trigger function is the way to go.
dbadmin.js - cleaned up the copy and deleteNodes functions

volunteers.js - bug fix - We were incorrectly writing a user's roles to his node.  We were writing a list of roles so the roles themselves were keyed using integer values.  What we need is for the children of 'roles' to be keyed by the role name with the value being string "true".  Not the greatest design/implementation, but that's what the rest of the app expects at this time.