Goofed by putting string values in users citizen_builder_id fields ("not populated yet").  This string value was causing the app to crash at User.login() for anyone that had this value because we were constructing a UserBean object from the DataSnapshot object.  And the UserBean object had typed the citizen_builder_id field as an Integer.
UserBean - so in this class, we changed the type of citizen_builder_id from Integer to String.
In setCitizen_builder_id() we changed the method signature to accept an Object.  Then inside this method, we examine what kind of Object we really have.
In getCitizen_builder_id(), we attempt to create a Long out of the citizen_builder_id field.  But if that fails (because the value is really "not populated yet"), then we return a -1

configure.js and userCreated.js - We make use of a new configuration parameter (on_citizen_builder_id) to determine whether we should really post user info to Slack or if we should do nothing.  This config parm just allows us to do some testing and data cleanup without sending a bunch of posts to slack that notifies everyone.