Bug fix in /videoListMain - We were ending the <table> after each row - oops.  Also, not displaying email content in the table anymore.  Don't want the markup in the emails to interfere with the markup that is displaying all the videos in /video/list
In deleteNodes, set a default limit on the number of nodes that will be deleted when no "limit" request parameter is supplied.  I was trying to delete all the /video/video_event nodes in one shot and firebase was complaining that too many triggers would fire if this were allowed.