Created a json() function and a readjson() function that are kind of hack/temporary functions which let me copy data from the dev database over to the prod database.  I didn't back up the /states node before updating/importing from OpenStates.  I really wanted that data backed up in case I needed to restore it or just figure out which elected officials are new in 2018 vs 2017.
Updated the copy() function to allow attribute/value pair to be passed in.  This is primarily because very big nodes can't just be copied with one action.  There's a 2GB memory limit that you run into on really big nodes.  You have to find an attribute and all its distinct values and copy a little over at a time - big pain.

Updated loadDivisionsTrigger() - still had references to the 'event' object which was a holdover from the Node6 version.  We're now in Node8.