restoreVideoCreator() - discovered that the Video Creator role could never be removed with the function written the way it was.  Modified it so that you could delete the role if the value was "false".  So to delete the Video Creator role, first set its value to "false" and then delete it.
userCreate(): When a user account is first created, if the user has satisfied legal, assign the (TelePatriot) Volunteer role to the person.  Streamlines onboarding because they don't have the TelePatriot Volunteer role in CB when the account is first created.  As result, the new user has to sign out and sign back in.  Doesn't sound like a big deal but you have to remember to do that.  Got several calls from Virginia about users who downloaded the app and couldn't see "My Mission" even though they had the TelePatriot Volunteer role in CB.  They couldn't see "My Mission" because when the TelePatriot Volunteer role is assigned in CB, there is no webhook or callback from CB to TP to tell TP that a role has been added.

With this change in place, the new user will have the (TelePatriot) Volunteer role applied to his account.  But the role will be removed on subsequent logins if the TelePatriot Volunteer role is not added to his account in CB.  This fix is meant to be a one-time/initialization fix to prevent confusion for the users and leadership.