Finally got the connection process pretty well stabilized.  Now, when the browser-based client loads the page, the connection doesn't happen right away.  That's because we commented out the connection request that the browser-based client used to make in the TelePatriot-Web project (video-invitation.component.js: ngOnInit())
So what did we do in THIS project...
VideoNode - getRemoteParticipant() and getRemoteParticipantUid().  These methods are used by VidyoChatFragment to tell if the browser-based client has loaded the page yet.
VideoParticipant - added a new attribute 'page_loaded'.  Gets set to true when the browser-based client loads the page
VidyoChatFragment - added new method isRemoteReady() that gets called from figureOutConnectivity().  If it turns out the remote participant's "page_loaded" attribute is true, then we call inviteLinks() to set the text in the app client to let the the app client know that the browser client has loaded the page and therefore is ready to connect.  So the app tells the user to press the green call button.  And then the connection is created according to the logic that's always been there - via a "connect request" VideoEvent